+++ package.json
@@ -59,6 +59,7 @@
     ]
   },
   "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@svgr/webpack": "^7.0.0",
     "@testing-library/react": "^14.0.0",
     "jest": "^27.5.1",
